video_generator/requirements.txt
# Gemini Video Generator - Requirements
# ======================================
# High-quality video generation with Gemini AI

# Core AI
google-generativeai>=0.8.0

# Video Processing
moviepy>=1.0.3
imageio>=2.34.0
imageio-ffmpeg>=0.4.9

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Audio Processing
pydub>=0.25.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Optional: For better font rendering
# fonttools>=4.40.0
